+++ burnman/__init__.py
@@ -90,4 +90,5 @@ from . import geotherm
 # miscellaneous
 from . import tools
 from . import nonlinear_fitting
+from . import eos_fitting
 from .partitioning import calculate_partition_coefficient, calculate_phase_percents
